BankingApp/src/test/java/io/github/bharatmane/banking/services/BankingServiceTest.java
package io.github.bharatmane.banking.services;

import io.github.bharatmane.banking.Prompter;
import io.github.bharatmane.banking.entity.Customer;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import java.io.ByteArrayOutputStream;
import java.io.PrintStream;
import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;
import java.util.Scanner;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

public class BankingServiceTest {
    private ByteArrayOutputStream outputStream;
    private PrintStream printStream;
    private Prompter prompter;
    private List<Customer> customers;


    @BeforeEach
    public void init() {
        outputStream = new ByteArrayOutputStream();
        printStream = new PrintStream(outputStream);
        customers = new ArrayList<Customer>(Arrays.asList(
                new Customer("jack","jack@123"),
                new Customer("jill","jill@123")
        ));
        Scanner scanner = new Scanner("1\njack\njack@123\n");
        prompter = new Prompter(printStream,scanner);

    }


    @Test
    @DisplayName("Should have valid list of customers when initialized")
    void  shouldHaveValidListOfCustomersWhenInitialized(){

        //Given
        BankingService bankingService = new BankingService(customers, prompter);

        //When

        //Then
        assertThat(bankingService.getCustomerCount()).isEqualTo(2);
    }

    @Test
    @DisplayName("Should show accounts menu when logged in with correct credentials")
    void  shouldShowAccountsMenuWhenLoggedInWithCorrectCredentials(){

        //Given
        BankingService bankingService = new BankingService(customers,prompter);

        //When
        bankingService.greet();

        //Then
        assertThat(outputStream.toString()).contains("1. Withdrawal");
    }

}
